==> prep/bd/spp/v2020/1_get_tet_iucn_spp.Rmd <==
---
title: 'OHI+ Tetiaroa: Species Trend'
author: 'Madeline Berger'
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3 
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/tet-prep/workflow/templates/tet_hdr.html'
  pdf_document:
    toc: true
---

# Summary 

This script compares the shapefile of Tetiaroa and its surrounding waters to the species range map shapefiles from the IUCN to extract all the species in this area. Methods match `1_get_ne_iucn_spp.Rmd` by Jamie Montgomery, OHI Northeast, and also use data and code developed by Casey O'Hara: https://github.com/madelinemberger/spp_risk_dists.

**Note:** For the species list pull, we decided to use a shapefile with a 50 nm buffer rather than a 3nm buffer. This is because when using the 3nm buffer, the species list was strangely short, and missing large groups that we know exist in Tetiaroa, like corals. Using the 50nm buffer helped capture these marine groups better, but also may have added certain species that do not exist in Tetiaroa as the 50nm buffer overlaps with Moorea and Tahiti. Ultimately the risk of adding extra species seemed less of an issue than missing all the coral species, so the layer created using the 50nm is the one used in the toolbox. 

# Set up

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warning = FALSE,
                      message = FALSE)

library(sf)
library(raster)
library(rfishbase) #included in ne, may not need
library(rredlist)
library(taxize)
library(here)
library(tidyverse)

source(here('workflow/R/common.R'))

#file paths
goal     <- 'spp'
scenario <- 'v2020'
#dir_anx       <- file.path(dir_M, 'git-annex/globalprep') 
dir_goal      <- file.path(here('prep/bd',goal, scenario))
dir_goal_anx  <- file.path(dir_anx, goal, scenario)
dir_github    <- '~/github/tet-prep'

#tetiaroa region shape file
tet_shp <- tet_buffer_3nm

tet_shp_50 <- tet_buffer_50nm

# api keys for IUCN 

api_file <- file.path(dir_M, 'git-annex/globalprep/spp', 
                      'api_key_gc.csv')
api_key <- scan(api_file, what = 'character')

```

# Methods

First step is to get a list of the cell IDs contained in the tet_region

```{r}

#these are global cell ids in raster format pulled from the spp_risk_dists repo.
cells <- raster("~/github/spp_risk_dists/_spatial/cell_id_rast.tif") #resolution is 10 km - is this too big? the area of this shp is 259 km2

#reproject shapefile to the same crs as the raster, which is WGS84

tet_reproj <- st_transform(tet_shp, crs = "+proj=cea +lon_0=0 +lat_ts=45 +x_0=0 +y_0=0 +ellps=WGS84
+units=m +no_defs")

tet_reproj_50 <- st_transform(tet_shp_50, crs = "+proj=cea +lon_0=0 +lat_ts=45 +x_0=0 +y_0=0 +ellps=WGS84
+units=m +no_defs")

plot(cells)
plot(tet_reproj_50, add = T)

#use extract to get the region ids for just the Tetiaroa area

tet_cells <- raster::extract(cells, tet_reproj) %>% 
  unlist() #this creates a vector out of the list from extract

tet_cells_50 <- raster::extract(cells, tet_reproj_50) %>% 
  unlist()

```

Read in files that contain all the species ranges and information

```{r}

# .csv file that lists all species and the file path to their map
spp_maps <- read_csv('~/github/spp_risk_dists/_data/spp_marine_maps_2019-2.csv',
                     col_types = 'ddciccc')

#file with species information to link to iucn_sid at the end
spp_info <- read_csv("/home/ohara/git-annex/spp_risk_dists/iucn/spp_info_from_api_2019-2.csv")

```

The following forloop goes through each species:  
- finds it's species path
- reads in the species .csv that lists all global cells where it is found
- filters that list to just those in the Northeast
- returns an aggregated dataframe (`taxa_cells_df`) with all species scientific names, their unique ids (sids), and cellIDs

```{r, message = FALSE}

#grab each taxa's folder filepath
taxa <- spp_maps$dbf_file %>% 
  unique() %>% 
  str_replace('\\....$', '')

#create an empty list that is the length of all taxa. We are going to fill this list  
taxa_cells_list <- vector('list', length = length(taxa))

taxa_cells_list_50 <- vector('list', length = length(taxa))
#for each taxa, grab the species map (raster) and filter to only keep those cells in the northeast.
for(i in seq_along(taxa)) { ### i <- 5
    taxon <- taxa[i]
    print(i)
    spp_ids_in_taxon <- spp_maps %>%
      filter(str_detect(dbf_file, taxon)) %>%
      .$iucn_sid
    cat(sprintf('processing %s spp in %s...\n', length(spp_ids_in_taxon), taxon)) #help with this, message?
    
    spp_cells <- lapply(spp_ids_in_taxon,
                                    FUN = function(x) { ### x <- spp_ids_in_taxon[1]
                                      f <- file.path('/home/ohara/git-annex/spp_risk_dists/spp_rasters_2019',
                                                     sprintf('iucn_sid_%s.csv', x))
                                      if(file.exists(f)) {
                                        y <- read_csv(f, col_types = 'di') %>%
                                          mutate(iucn_sid = x) %>%
                                          dplyr::select(-presence)  %>%
                                          filter(cell_id %in% tet_cells_50)
                                      } else {
                                        y <- data.frame(cell_id = NA,
                                                        iucn_sid = x, 
                                                        f = f, error = 'file not found')
                                      }
                                      return(y)
                                    }) %>%
      bind_rows() %>%
      mutate(spp_gp = taxon)
    
    taxa_cells_list_50[[i]] <- spp_cells
}

taxa_cells_list[[3]] #was using to just look at a few
taxa_cells_df_50 <- taxa_cells_list_50 %>%
    bind_rows() %>% 
    filter(!is.na(cell_id)) %>% 
    dplyr::select(iucn_sid) %>% 
    distinct() %>%
    left_join(spp_info)  
```

Small note on looking at species maps - use filepath?
```{r}

humpback_whale <- spp_maps %>% 
  filter(iucn_sid == 13006)

```

Attach common names

```{r}
#create list of scintific names of species

scinames <- taxa_cells_df_50$sciname

#create loop around redlist function rl_search to find common names  
#create empty data frame that will be filled
iucn_common_names_50 <- data.frame() 

for(i in 1:length(scinames)){
  sp <- scinames[i]
  print(i)
  if(sp %in% c("Procelsterna cerulea")){ #this one threw some error and no $result was returned from rl_search
  comm <- NA
  }else{
  comm <- rredlist::rl_search(sp, key = api_key )$result$main_common_name
  }
  
  df <- data.frame(sciname = sp,
                   common = comm)
  
  iucn_common_names_50 <- rbind(iucn_common_names_50, df) %>%
    mutate(common = as.character(common),
         sciname = as.character(sciname))
}

#had to remove blue noddy because it was causing an error

```

Check out the ones we are still missing for common name

```{r}

miss_sp_50 <- iucn_common_names_50 %>%
  filter(is.na(common))

#for 50nm there are 241 missing

```

We are still missing `r nrow(filter(iucn_common_names_50, is.na(common)))`. Use taxize package to try and find these.

```{r}

#create vector from missing scinames
missing_scinames_50 <- iucn_common_names_50%>%
  filter(is.na(common)) %>%
  .$sciname

#loop to pull out the common names from taxize

common_names <- c()
for(i in 1:length(missing_scinames_50)){ #i <- 1
  sp <- missing_scinames_50[i]
  comm <- taxize::sci2comm(sp, db = "itis")
  print(comm)
 
  common_names <- c(common_names, comm)


}

#output is list, turn into dataframe

df <- do.call(rbind,lapply(common_names,data.frame)) #this isn't running for me
df$sciname <- rownames(df)

```

This only got us blue noddy, not really helpful. Try some exploration

```{r}
#this is a sea cucumber, does not have a common name in IUCN
rl_search("Holothuria olivacea", key = api_key )$result$main_common_name

taxize::get_wormsid("Holothuria olivacea")

scinames_explore <- miss_sp %>%
  separate(sciname, sep = " ", into = c("genus", "family")) %>% 
  group_by(genus) %>% 
  summarize(
    count = length(family)
  )


```

There are only 6 genus' represented in the missing data: 
- 8 of them are Holothuria, which are sea cucumbers
- `scinames_explore$genus[1]` are slipper lobsters
- `scinames_explore$genus[3]` are also sea cucumbers
- `scinames_explore$genus[4]` mostly deep sea blind lobsters
- `scinames_explore$genus[5]` these are the noddy's will add in manually below
- `scinames_explore$genus[6]` more lobsters?

```{r}

spp_names <- iucn_common_names %>% 
  mutate(common_name = case_when(
              sciname == "Procelsterna cerulea" ~ "Blue Noddy",
              TRUE ~ as.character(common)))

#add in the rest of the info 

iucn_spp_info <- taxa_cells_df %>%
  select(iucn_sid, sciname, population, category) %>% 
  left_join(., spp_names, by = "sciname") %>% 
  select(-common) %>% 
  distinct()

iucn_spp_info_50 <- taxa_cells_df_50 %>% 
  dplyr::select(iucn_sid, sciname, population, category) %>% 
  left_join(., iucn_common_names_50, by = 'sciname') %>% 
  distinct()

```

Write out csv to folder `int`

```{r}

#save
write_csv(iucn_spp_info, "int/1_iucn_spp_in_tet.csv")

#using 50 nm buffer
write_csv(iucn_spp_info_50, "int/1_iucn_spp_in_tet_50.csv")

```

Last step is to get a list of the maps now that we have the names before moving onto to script 2
(come back to this since the subpop region is tricky with this region)

Note: the marine maps df only has 128 rows, meaning we are missing 5 maps
```{r}
# only loggerhead sea turtle has subpops so going to filter out the one we want
spp_marine_maps <- read_csv("~/github/spp_risk_dists/_data/spp_marine_maps_2018-1.csv") %>%
  filter(sciname %in% iucn_spp_info_50$sciname, 
         subpop %in% c("Pacific South", NA))

write.csv(spp_marine_maps, "int/1_iucn_spp_shp_filepaths_50.csv")

```


Double check the ICO list - most should be included now

1. Get files needed 
```{r}

#read in ico species list

ico_sp <- read_csv(here("/prep/sp/ico/v2020/int", "tet_ico_all.csv"))

spp_sp <- read_csv(here("/prep/bd/spp/v2020/int","1_iucn_spp_in_tet_50.csv")) %>% 
  mutate(
    sciname = str_to_lower(sciname),
    name = str_to_lower(common)
  ) %>% 
  dplyr::select(-common, name, sciname, iucn_sid, population, category)

```

We can detect any differences using `setdiff()`?

```{r}
#this is a list of whats in ico but NOT in spp - this is what we need to add

ico_not_sp <- setdiff(ico_sp$sciname, spp_sp$sciname) 

#this is a list of whats in spp but NOT ICO, just as a reference
setdiff(spp_sp$sciname, ico_sp$sciname) 

```

So still missing 21 - need to add these back in 

Filter the ico list for the missing ones and bind to the iucn list

```{r}

spp_all <- ico_sp %>% 
  filter(sciname %in% ico_not_sp) %>% 
  bind_rows(., spp_sp) #create new dataframe with all the species

#write this out 

write_csv(spp_all, file.path(dir_goal, "/int/2_spp_ico_iucn_50.csv"))

```


Get differences between 50nm pull and 3nm pull
Post meeting deliverable - difference between 3 nm and 50 nm

```{r}

#read in he data sets

tet_3nm <- read_csv(file.path(dir_goal, "int", "2_spp_ico_iucn.csv"))

tet_50nm <- read_csv(file.path(dir_goal, "int", "2_spp_ico_iucn_50.csv"))


first <- setdiff(tet_3nm$sciname, tet_50nm$sciname)

second <- setdiff(tet_50nm$sciname, tet_3nm$sciname)

tet_additional <- tet_50nm %>%
  filter(sciname %in% second)

write_csv(tet_additional, file.path(dir_goal, "int", "difference_3_50.csv"))
  
  

```

***
