==> prep/bd/spp/v2020/2_get_spp_scores.Rmd <==
---
title: "OHI+ Tetiaroa: Get species scores"
author: "Madeline Berger"
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3 
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/tet-prep/workflow/templates/tet_hdr.html'
  pdf_document:
    toc: true
---

# Summary 

This script joins the different conservation status to their respective weights, in addition to creating a trends layer using the `rl_search()` function in the rredlist package

# Set up

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warning = FALSE,
                      message = FALSE,
                      eval = FALSE)

library(rredlist)
library(here)
library(tidyverse)

source(here('workflow/R/common.R'))

#file paths
goal     <- 'spp'
scenario <- 'v2020'
dir_anx       <- file.path(dir_M, 'git-annex/globalprep') 
dir_goal      <- file.path(here::here('prep/bd',goal, scenario))
dir_goal_anx  <- file.path(dir_anx, goal, scenario)
dir_github    <- '~/github/tet-prep'


# api keys for IUCN 

api_file <- file.path(dir_M, 'git-annex/globalprep/spp', 
                      'api_key_gc.csv')
api_key <- scan(api_file, what = 'character')

```

# Status

Read in species list

```{r}

spp_list <- read_csv(here::here("/prep/bd/spp/v2020/int", "2_spp_ico_iucn_50.csv")) %>%
      dplyr::select(name, sciname, iucn_sid, category) %>%
      distinct()


```

First, just score this list

```{r}

pop_cat <- data.frame(category = c("LC", "NT", "VU", "EN", "CR", "EX", "T", "CD", "NE", "DD"), 
                      risk_score = c(   0,  0.2,  0.4,  0.6,  0.8,  1.0, 0.6,  0.3,   NA,  NA),
                      stringsAsFactors = FALSE) %>% 
  mutate(
    stat_score = 1 - risk_score
  )

#join the scores data to the list data - note, this doesn't have year. 

spp_status <- left_join(spp_list, pop_cat, by = "category") %>%
  mutate(stat_score = case_when(
    #category %in% c("LR/lc") ~ 1.0, #this one isn't showing up? double check 
   # category %in% c("LR/nt") ~ 0.8,
    category %in% c("LR/cd") ~ 0.9,
    TRUE ~ stat_score),
   year = 2020) %>% 
  dplyr::select(-risk_score)

#double check that this is what the layer should look like 

write.csv(spp_status, file.path(dir_goal,"int","3_iucn_status_scores.csv"), row.names = F)
#write.csv(spp_status, file = "~/github/tet-scores/region/layers/spp_status.csv", row.names = F)

```

# Trend
Next, get trends using the method Jamie does
```{r, echo=FALSE}

#forloop for each species to grab trend

df_iucn <- data.frame()
for(i in 1:nrow(spp_list)){
  
  print(i)
  sp <- as.character(spp_list[i,2]) #grab scientific name
  
  tr <- rl_search(sp, key = api_key)$result$population_trend
  
  if(is.null(tr)){
  
  df2 <- data.frame(sciname = sp,
                    trend = NA)
  }else{
    df2 <- data.frame(sciname = sp,
                      trend = tr)
  }
  
  df_iucn <- rbind(df_iucn, df2)
}

#write out to int folder

write.csv(df_iucn, "3_iucn_population_trends.csv")
```

## Score
Now, score the trend based on what was pulled from rl_search
```{r}

iucn_trends <- read_csv(here::here("prep/bd/spp/v2020/int", "3_iucn_population_trends.csv"))

#score table
pop_trend_score <- data.frame(trend = c("Increasing", "Decreasing", "Stable", "Unknown"),
                              score = c(0.025, -0.025, 0, NA))
```

Join to the trend list

```{r}

spp_trend_scores <- iucn_trends %>% 
  left_join(pop_trend_score) %>%
  filter(!is.na(trend)) #filter out three where trend was unknown
```

## Visualize

```{r}
spp_trends_plot <- spp_trend_scores %>%
  group_by(trend) %>%
  summarize(count = n())

spp_trends_plot

ggplot(spp_trends_plot, aes(x = trend, y = count)) +
  geom_bar(stat = "identity") +
  theme_bw()

#over half of them are unknown

```

##Save
```{r}

spp_trends <- spp_trend_scores %>%
  mutate(year = 2020) %>%   #have to add a year for the toolbox - should this be 2020
  dplyr::select(sciname, score, year)

write.csv(spp_trends, file = "~/github/tet-scores/region/layers/spp_trend.csv", row.names = F)  #save to toolbox



```

# Averaging by class

To get a more representative score, we are going to average each score over class. This will avoid a class that has a lot of species but may not be that important, ie cone snails, from dominating the score. The averaging will be done in the toolbox, however here we'll add the class and then export that to the toolbox. 

```{r}
#get full list of species with taxonomic groups
out <- rl_sp(all = TRUE, key = api_key)

all_df <- do.call(rbind, lapply(out, "[[", "result")) #creates data frames from each item in the large list in `out` and then binds them

all_df_comp <- all_df %>%
  dplyr::select(-infra_rank, -infra_name) %>%
    dplyr::rename(iucn_sid = taxonid, sciname = scientific_name, class = class_name) %>% 
    setNames(names(.) %>%
               stringr::str_replace('_name', '')) %>% 
  mutate(
    sciname = str_to_lower(sciname),
    class = str_to_lower(class)
  ) %>% 
dplyr::select(class,sciname,population, category)

#read in our full spp_list - this is the 50 one

spp_list <- read_csv(file.path(dir_goal,"int","3_iucn_status_scores.csv"))

#filter large df by the all the scientific names vector

spp_class <- inner_join(spp_list, all_df_comp, by = "sciname")


```

Now need to filter out the sub-populations that I want, since their conservation status is different. This is a bit confusing because there are some species where the specific subpop is relevant, and others that we don't want the subpop for. I removed all the subpopulations, leaving only "NA" in the df, manually grabbed the ones I wanted from that, and then re-joined those into the larger df. 

```{r}

subpops <- spp_class %>% 
  filter(!is.na(population))
subpops_nm <- unique(subpops$sciname) #16 subpops, only 6 we want the rest can be NA

test <- spp_class[grepl("South Pacific", spp_class$population) | grepl("Hawaiian", spp_class$population) | grepl("Asia", spp_class$population)| grepl("East Pacific", spp_class$population) | grepl("Indo-west", spp_class$population) | grepl("Oceania", spp_class$population),] #these are the ones I want

spp_status_final <- spp_class %>% filter(is.na(population)) %>% 
  filter(sciname %!in% test$sciname) %>% 
  bind_rows(., test) %>%
  mutate(region_id = 1) %>% 
  mutate(
    stat_score = case_when(
      sciname %in% "birgus latro" ~ 0.4, #because they are locally endangered
      TRUE ~ stat_score
    )
  ) %>% 
  dplyr::select(region_id,common = name,class,sciname,category = category.x, status = stat_score, year)

#write spp_stats to outputs and toolbox

write.csv(spp_status_final, file.path("~/github/tet-scores/region/layers/spp_status.csv"), row.names = F)

write.csv(spp_status_final, file.path(dir_goal,"output","spp_status.csv"), row.names = F)
```

Final trends layer is created in the last script `3_get_historic_trends.Rmd`.
