==> prep/bd/spp/v2020/3_get_historic_trends.Rmd <==
---
title: "OHI+ Tetiaroa: Finding Species Trends"
author: "Madeline Berger"
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3 
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/tet-prep/workflow/templates/tet_hdr.html'
  pdf_document:
    toc: true
---

## Summary 

Script based on code developed by Gage Clawson to try and calculate species trend using historical assessments and trend models. 

## Set up 
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warning = FALSE,
                      message = FALSE,
                      eval = FALSE)

library(rredlist)
library(tidyverse)
library(dplyr)
library(rgdal)
library(raster)
library(sf)
library(here)
library(ggridges)
library(ggplot2)
library(stringr)
library(plyr)

source(here::here('workflow/R/common.R'))

#### functions

"%!in%" <- function(x,y)!('%in%'(x,y))

cat_msg <- function(x, ...) { 
  if(is.null(knitr:::.knitEnv$input.dir)) {    ### not in knitr environment, so use cat()    
  cat(x, ..., '\n')  } 
  else {    ### in knitr env, so use message()    
  message(x, ...)  }  
  return(invisible(NULL))
  
  }

#### api keys for rredlist

api_file <- file.path(dir_M, 'git-annex/globalprep/spp', 
                      'api_key_gc.csv')
api_key <- scan(api_file, what = 'character')  

api_version <- '2020-1'

## file paths 
goal     <- 'spp'
scenario <- 'v2020'
dir_anx       <- file.path(dir_M, 'git-annex/globalprep') 
dir_goal      <- file.path(here::here('prep/bd',goal, scenario))
dir_goal_anx  <- file.path(dir_anx, goal, scenario)
dir_github    <- '~/github/tet-prep'


```


## Method 

read in data - use what I already had from `2_get_spp_scores.csv`
```{r, eval = FALSE}
## Load packages and other important things
tet_pop_trends <- read_csv(here::here("/prep/bd/spp/v2020/int", "3_iucn_population_trends.csv"))

tet_species_vec <- tet_pop_trends$sciname
```


``` {r, eval = FALSE}
## get past assessments with rl_history
hist <- list()


## This for loop breaks sometimes, so just re run it and itll continue where it left off (because of the if statement I added)
for(i in tet_species_vec){
#i = “Sula leucogaster”
if(i %!in% names(hist)){
      cat_msg("Getting species history information for this species:", i)
spp_hist_temp <- rl_history(name = i, key = api_key) # get a species history
spp_hist_temp2 <- data.frame(cbind(sciname = spp_hist_temp$name, spp_hist_temp$result)) #combine into one dataset
hist[[i]] <- spp_hist_temp2
} else {
    cat_msg("File of API species history exists in list, skipping this: \n", i)
}

}

#get the list into a df

spp_hist <- ldply(hist, rbind)

#write out csv so you can run loop later   

write_csv(spp_hist, file.path(dir_goal, "int/spp_history.csv"))
```


Join to categories 
``` {r, eval = FALSE}


cat_lookup <- read_csv(file.path(dir_goal, "risk_code_lookup.csv")) # read in the category look up so that old category naming conventions will match todays. This is in ohiprep_v2020/globalprep/spp/v2020/setup/raw folder

spp_hist <- read_csv(file.path(dir_goal, "int/spp_history.csv"))
    
spp_past_df <- spp_hist %>%
    left_join(cat_lookup, by = c('code', 'category')) %>%
    dplyr::rename(
           old_cat  = code,
           cat_txt  = category,
           cat_ts   = code_current) #rename stuff and get scores

pop_cat <- data.frame(cat_ts       = c('LC', 'NT', 'VU', 'EN', 'CR', 'EX', 'NE', 'DD'),
                        cat_ts_score = c( 0.0,  0.2,  0.4,  0.6,  0.8,  1.0,  NA,   NA)) ## define scores again

spp_past_df1 <- spp_past_df %>%
    left_join(pop_cat, by = 'cat_ts') %>%
    filter(!is.na(cat_ts_score) & !is.na(year)) %>%
    arrange(sciname, year) %>%
    dplyr::select(sciname, year, cat_ts, cat_ts_score) %>%
    mutate(iucn_version = api_version) #this tells you what year it is pulling from 

```


Create data frames for modeling 

- `cat_trend` binds our historic pull to the trend, so contains all info for all years
- `lm_trend` filters out only ones that have enough info to create a linear model on 

``` {r, eval = FALSE}
tet_pop_trends <- read_csv(here::here("/prep/bd/spp/v2020/int", "3_iucn_population_trends.csv")) %>%
  dplyr::rename(pop_trend = trend) ## read in population trends grabbed from IUCN

cat_trend <- tet_pop_trends %>%
  left_join(spp_past_df1, by = "sciname") %>%
   group_by(sciname) %>%
  dplyr::mutate(n_assess = n()) %>%
  ungroup()
cat_trend$year <- as.numeric(cat_trend$year) #this df has ALL THE INFO

lm_trend <- cat_trend %>%
  filter(n_assess >= 2) %>% #only include ones with enough assessments
  group_by(sciname) %>%
  do(calc_trend = lm(cat_ts_score ~ year, data = .)[['coefficients']][['year']]) %>%
  mutate(calc_trend = as.numeric(calc_trend)) %>%
  mutate(calc_trend = round(calc_trend, 5)) ## Run a linear regression to get trend based on historical assessments
```

Compare IUCN trend to what we found with model
``` {r, eval = FALSE}
trend <- cat_trend %>%
  dplyr::filter(n_assess > 1) %>%
  dplyr::group_by(sciname) %>%
  dplyr::summarize(years  = paste(year, collapse = ', '),
            cat_ts = paste(cat_ts, collapse = ', '),
            scores = paste(cat_ts_score, collapse = ', '),
            pop_trend_desc = first(pop_trend)) %>%
  ungroup() %>%
  left_join(lm_trend, by = c('sciname'))

trend_2 <- trend %>%
  filter(!is.na(pop_trend_desc)) %>%
  mutate(pop_trend = case_when(pop_trend_desc == 'Decreasing' ~ -1,
                               pop_trend_desc == 'Increasing' ~  1,
                               pop_trend_desc == 'Stable'     ~  0,
                               TRUE                           ~ NA_real_)
         )

```

Next model: taking `calc_trend` and modeling it on the IUCN population trend 

``` {r, eval = FALSE}  
  ### NOTE: here we’re coding the trend in *population* not risk. This tells us the regression coeffs based on the population trends grabbed from iucn (increasing, decreasing, unknown)

trend_regr <- lm(calc_trend ~ pop_trend, data = trend_2)
trend_coeffs <- summary(trend_regr)
```

``` {r}
### manual values from linear regression:
slope <- trend_coeffs$coefficients[1] %>% round(5) #grabs slope from model
trend_score_lookup <- c('Decreasing' = -slope, 'stable' = 0, 'Increasing' = slope) #this is for gapfilling if the model didn't find anything

trend <- trend_2 %>%
  dplyr::select(sciname, calc_trend) %>%
  distinct()

trend_df_2 <- tet_pop_trends %>%
  left_join(spp_past_df1, by = c('sciname')) %>%
  distinct() %>%
  left_join(trend, by = c('sciname')) %>%
  mutate(trend_score  = ifelse(is.na(calc_trend), trend_score_lookup[pop_trend], calc_trend),
         trend_source = case_when(!is.na(calc_trend)  ~ 'lm',
                                  !is.na(trend_score) ~ 'regr',
                                  TRUE ~ NA_character_)) %>%
  dplyr::select(sciname,  pop_trend, trend_score, trend_source) %>%
  distinct() ## this is our final trend data set


write_csv(trend_df_2, file.path(dir_goal, "int/spp_hist_trends.csv"))
```

Last step is to fill in the ones that have a pop_trend but were NA in the loop, meaning they probably only had one observation.

```{r, eval = FALSE}

tet_pop_trends <- read_csv(here::here("/prep/bd/spp/v2020/int", "3_iucn_population_trends.csv")) %>%
  dplyr::rename(pop_trend = trend)

trends <- read_csv(here::here("prep/bd/spp/v2020/int", "spp_hist_trends.csv"))

lm_trends <- trends %>% 
  filter(!is.na(trend_score))

#quickly add class
class_df <- read_csv(file.path(dir_goal, "int", "class_df.csv"))

#filter out the ones that have a pop_trend but no calc_trend, then re add with new score

all_trends <- trends %>% 
  filter(is.na(trend_score)) %>%
  mutate(
    trend_score = case_when(
      pop_trend %in% "Stable" ~ 0,
      TRUE ~ trend_score
    ),
    trend_source = "iucn"
  ) %>% 
  bind_rows(., lm_trends) %>% 
  inner_join(., class_df, by = "sciname") %>% 
  mutate(
    region_id = 1, 
    year = 2020
    )#yay


#write out final trend scores to output and to toolbox

write_csv(all_trends, file.path(dir_goal, "output/spp_trends.csv"))

write.csv(all_trends %>% dplyr::select(region_id, sciname, class, trend = trend_score, trend_source, year), file = "~/github/tet-scores/region/layers/spp_trend.csv", row.names = F)

```
